Moving LibraryListViewHeader to common
